+++ CHANGELOG.md
@@ -25,3 +25,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Make Java Parameters entry larger
 - Remove servers button from system instances
 - Move Clone button on instance to right click menu
+- Remove the fabric api install button after it's installed